requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.24.2
botocore==1.27.2
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
jmespath==1.0.0
numpy==1.22.4
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.2
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.19.0
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.10.0
urllib3==1.26.9
webdriver-manager==3.7.0
wsproto==1.1.0
